Share reduced motion media query hook
